Moved Gui code to Gui package
